02-MarketPulse-Agent/requirements.txt
langchain
langchain-groq
langchain-community
tavily-python
python-dotenv
streamlit